LPS-136032 Revert LPS-135504 automatic backwards compatibilty.
This reverts commits: e3e5fcd439b0870ec38174955d015eae4f8f7267, 01ade2ebe79f3d524cd24c9ec422795a1a778297, and 035e66e1eafcb83509ca08795d493015c7a01709.
